requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
google-generativeai>=0.3.2
python-dotenv>=1.0.1
plotly==5.19.0
protobuf>=4.25.3,<5.0
openai>=1.0.0
transformers>=4.30.0 